+++ extension/content/panel.js
@@ -47,7 +47,7 @@
   };
 
   var shadowRoot = null;
-  var APP_VERSION = '0.1.0';
+  var APP_VERSION = '0.2.0';
   var APP_URL = 'https://wowmd.app/';
 
   function ensureShadowRoot() {